backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
biopandas==0.5.1
biopython==1.85
blinker==1.9.0
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
dotenv==0.9.9
enum34==1.1.10
exceptiongroup==1.2.2
fastapi==0.115.12
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.57.0
groq==0.24.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.9.4
meeko==0.6.1
numpy==1.26.4
packaging==25.0
pandas==2.2.3
pillow==11.2.1
py3Dmol==2.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
typing==3.7.4.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
xmltramp2==3.1.1
zipp==3.21.0
gunicorn==20.1.0
ramplot==1.0.4
MDAnalysis==2.7.0
psycopg2-binary>=2.9.5